+++ src/components/super/i-block/state/index.ts
@@ -17,6 +17,7 @@ import type Async from 'core/async';
 import type { BoundFn } from 'core/async';
 
 import { initGlobalEnv } from 'core/env';
+import { i18nFactory } from 'core/prelude/i18n';
 import { component, remoteState, hook, hydrationStore, Hook, State } from 'core/component';
 
 import type bRouter from 'components/base/b-router/b-router';
@@ -65,7 +66,7 @@ export default abstract class iBlockState extends iBlockMods {
 	@computed({watchable: true})
 	get remoteState(): State {
 		if (SSR) {
-			return {...this.ssrState, ...remoteState};
+			return {...this.ssrState!, ...remoteState};
 		}
 
 		return remoteState;
@@ -288,6 +289,35 @@ export default abstract class iBlockState extends iBlockMods {
 		}
 	}
 
+	/**
+	 * Factory for creating internationalizing function
+	 */
+	i18n(
+		keysetName: CanArray<string>, customLocale?: Language
+	): (key: string | TemplateStringsArray, params?: I18nParams) => string {
+		return i18nFactory(keysetName, customLocale ?? this.remoteState.lang);
+	}
+
+	/**
+	 * A function for internationalizing texts inside traits.
+	 * Because traits are called within the context of components, standard `i18n` does not work,
+	 * and you need to explicitly pass the key set name (trait names).
+	 *
+	 * @param traitName - the trait name
+	 * @param text - the text for internationalization
+	 * @param [opts] - additional internationalization options
+	 */
+	i18nTrait(traitName: string, text: string, opts?: I18nParams): string {
+		return this.i18n(traitName)(text, opts);
+	}
+
+	/**
+	 * A function for internationalizing texts
+	 */
+	get t(): (key: string | TemplateStringsArray, params?: I18nParams) => string {
+		return this.i18n(this.componentI18nKeysets);
+	}
+
 	/** @inheritDoc */
 	getComponentInfo(): Dictionary {
 		return {
@@ -466,6 +496,7 @@ export default abstract class iBlockState extends iBlockMods {
 		const
 			i = this.instance;
 
+		this.i18n = i.i18n.bind(this);
 		this.syncStorageState = i.syncStorageState.bind(this);
 		this.syncRouterState = i.syncRouterState.bind(this);
 		this.initGlobalEnv = i.initGlobalEnv.bind(this);